refactor: 상수 파일 분리 + 사용하지 않는 코드 제거 (#131)
