--- src/app/admin/test/_constants/index.ts
@@ -1,1 +1,0 @@
-export * from './testListLabels'
